requirements.txt
channels==4.0.0
Django==4.1
djangorestframework==3.14.0
PyMySQL==1.1.0
